Fix Windows shared

+++ conanfile.py
@@ -39,19 +39,19 @@ class LibmikmodConan(ConanFile):
     _build_subfolder = "build_subfolder"
 
     def config_options(self):
-        if self.settings.os == 'Windows':
+        if self.settings.os == "Windows":
             del self.options.fPIC
         else:
             del self.options.with_dsound
             del self.options.with_mmsound
-        if self.settings.os != 'Linux':
+        if self.settings.os != "Linux":
             del self.options.with_alsa
         # Non-Apple Unices
-        if self.settings.os not in ['Linux', 'FreeBSD']:
+        if self.settings.os not in ["Linux", "FreeBSD"]:
             del self.options.with_oss
             del self.options.with_pulse
         # Apple
-        if self.settings.os not in ['Macos', 'iOS', 'watchOS', 'tvOS']:
+        if self.settings.os not in ["Macos", "iOS", "watchOS", "tvOS"]:
             del self.options.with_coreaudio
 
     def configure(self):
@@ -61,9 +61,9 @@ class LibmikmodConan(ConanFile):
     def requirements(self):
         if self.settings.os == "Linux":
             if self.options.with_alsa:
-                self.requires('libalsa/1.1.9')
+                self.requires("libalsa/1.1.9")
             if self.options.with_pulse:
-                self.requires('pulseaudio/13.0@bincrafters/stable')
+                self.requires("pulseaudio/13.0@bincrafters/stable")
 
     def source(self):
         tools.get(**self.conan_data["sources"][self.version])
@@ -72,14 +72,14 @@ class LibmikmodConan(ConanFile):
 
     def _configure_cmake(self):
         cmake = CMake(self, set_cmake_flags=True)
-        cmake.definitions['ENABLE_STATIC'] = not self.options.shared
-        cmake.definitions['ENABLE_DOC'] = False
-        cmake.definitions['ENABLE_DSOUND'] = self._get_safe_bool('with_dsound')
-        cmake.definitions['ENABLE_MMSOUND'] = self._get_safe_bool('with_mmsound')
-        cmake.definitions['ENABLE_ALSA'] = self._get_safe_bool('with_alsa')
-        cmake.definitions['ENABLE_OSS'] = self._get_safe_bool('with_oss')
-        cmake.definitions['ENABLE_PULSE'] = self._get_safe_bool('with_pulse')
-        cmake.definitions['ENABLE_COREAUDIO'] = self._get_safe_bool('with_coreaudio')
+        cmake.definitions["ENABLE_STATIC"] = not self.options.shared
+        cmake.definitions["ENABLE_DOC"] = False
+        cmake.definitions["ENABLE_DSOUND"] = self._get_safe_bool("with_dsound")
+        cmake.definitions["ENABLE_MMSOUND"] = self._get_safe_bool("with_mmsound")
+        cmake.definitions["ENABLE_ALSA"] = self._get_safe_bool("with_alsa")
+        cmake.definitions["ENABLE_OSS"] = self._get_safe_bool("with_oss")
+        cmake.definitions["ENABLE_PULSE"] = self._get_safe_bool("with_pulse")
+        cmake.definitions["ENABLE_COREAUDIO"] = self._get_safe_bool("with_coreaudio")
         cmake.configure(build_folder=self._build_subfolder, source_folder=self._source_subfolder)
         return cmake
 
@@ -99,9 +99,9 @@ class LibmikmodConan(ConanFile):
                 tools.patch(**patch)
  
          # Ensure missing dependencies yields errors
-        tools.replace_in_file(os.path.join(self._source_subfolder, 'CMakeLists.txt'),
-                              'MESSAGE(WARNING',
-                              'MESSAGE(FATAL_ERROR')
+        tools.replace_in_file(os.path.join(self._source_subfolder, "CMakeLists.txt"),
+                              "MESSAGE(WARNING",
+                              "MESSAGE(FATAL_ERROR")
  
         tools.replace_in_file(os.path.join(self._source_subfolder, "drivers", "drv_alsa.c"),
                               "alsa_pcm_close(pcm_h);",
@@ -114,17 +114,19 @@ class LibmikmodConan(ConanFile):
         self.copy(pattern="COPYING.LESSER", dst="licenses", src=self._source_subfolder)
         cmake = self._configure_cmake()
         cmake.install()
-        tools.rmdir(os.path.join(self.package_folder, "bin"))
+        os.remove(os.path.join(self.package_folder, "bin", "libmikmod-config"))
+        if not self.options.shared:
+            tools.rmdir(os.path.join(self.package_folder, "bin"))
         tools.rmdir(os.path.join(self.package_folder, "lib", "pkgconfig"))
 
     def package_info(self):
         self.cpp_info.libs = tools.collect_libs(self)
         if not self.options.shared:
-            self.cpp_info.defines = ['MIKMOD_STATIC']
+            self.cpp_info.defines = ["MIKMOD_STATIC"]
 
-        if self._get_safe_bool('with_dsound'):
-            self.cpp_info.system_libs.append('dsound')
-        if self._get_safe_bool('with_mmsound'):
-            self.cpp_info.system_libs.append('winmm')
-        if self._get_safe_bool('with_coreaudio'):
-            self.cpp_info.frameworks.append('CoreAudio')
+        if self._get_safe_bool("with_dsound"):
+            self.cpp_info.system_libs.append("dsound")
+        if self._get_safe_bool("with_mmsound"):
+            self.cpp_info.system_libs.append("winmm")
+        if self._get_safe_bool("with_coreaudio"):
+            self.cpp_info.frameworks.append("CoreAudio")
